Fix automatic module name and split package issue
Gitlab issue: #207
Since Java 9 modularity, if JARs used in a modular environment doesn't
provide a module name, the JAR name will be used by default. The module
packetfactory-static caused an issue because of the static keyword so I
told maven to add the automatic module name in the manifest without the
static keywork.
Also, classes in separate modules must not have the same package path,
this is commonly called the split-package issue. There was a conflict
between the packetfactory-static and pcap4j-core because they both
declared classes in ord/pcap4j/factory. I changed the
packetfactory-static packages to avoid the splitpackage issue.
