+++ pcap4j-packetfactory-propertiesbased/pom.xml
@@ -88,6 +88,18 @@
           </systemPropertyVariables>
         </configuration>
       </plugin>
+      <!-- Provides a module name since Java 9 modularity -->
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-jar-plugin</artifactId>
+        <configuration>
+          <archive>
+            <manifestEntries>
+              <Automatic-Module-Name>org.pcap4j.packet.factory.propertiesbased</Automatic-Module-Name>
+            </manifestEntries>
+          </archive>
+        </configuration>
+      </plugin>
     </plugins>
   </build>
 </project>
