+++ pcap4j-packetfactory-static/pom.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!-- Copyright (C) 2011-2015  Pcap4J.org -->
+<!-- Copyright (C) 2011-2015 Pcap4J.org -->
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
@@ -73,6 +73,18 @@
               target/test
             </org.pcap4j.packet.AbstractPacketTest.tmpDir>
           </systemPropertyVariables>
+        </configuration>
+      </plugin>
+      <!-- Provides a module name since Java 9 modularity -->
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-jar-plugin</artifactId>
+        <configuration>
+          <archive>
+            <manifestEntries>
+              <Automatic-Module-Name>org.pcap4j.packet.factory.impl</Automatic-Module-Name>
+            </manifestEntries>
+          </archive>
         </configuration>
       </plugin>
     </plugins>
