+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/impl/AbstractStaticPacketFactory.java
@@ -5,14 +5,16 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.impl;
 
 import java.util.HashMap;
 import java.util.Map;
+
 import org.pcap4j.packet.IllegalPacket;
 import org.pcap4j.packet.IllegalRawDataException;
 import org.pcap4j.packet.Packet;
 import org.pcap4j.packet.UnknownPacket;
+import org.pcap4j.packet.factory.PacketFactory;
 import org.pcap4j.packet.namednumber.NamedNumber;
 
 /**
