+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/impl/PacketFactoryBinder.java
@@ -5,10 +5,11 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.impl;
 
 import java.util.HashMap;
 import java.util.Map;
+
 import org.pcap4j.packet.DnsResourceRecord.DnsRData;
 import org.pcap4j.packet.IcmpV6CommonPacket.IpV6NeighborDiscoveryOption;
 import org.pcap4j.packet.IpV4InternetTimestampOption;
@@ -22,6 +23,8 @@ import org.pcap4j.packet.Packet;
 import org.pcap4j.packet.RadiotapPacket.RadiotapData;
 import org.pcap4j.packet.SctpPacket.SctpChunk;
 import org.pcap4j.packet.TcpPacket.TcpOption;
+import org.pcap4j.packet.factory.PacketFactory;
+import org.pcap4j.packet.factory.StaticUnknownPacketFactory;
 import org.pcap4j.packet.namednumber.DataLinkType;
 import org.pcap4j.packet.namednumber.Dot11FrameType;
 import org.pcap4j.packet.namednumber.EtherType;
