+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/impl/StaticDataLinkTypePacketFactory.java
@@ -5,7 +5,7 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.impl;
 
 import org.pcap4j.packet.BsdLoopbackPacket;
 import org.pcap4j.packet.Dot11Selector;
