+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/impl/StaticDot11FrameTypePacketFactory.java
@@ -5,7 +5,7 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.impl;
 
 import org.pcap4j.packet.Dot11ProbeRequestPacket;
 import org.pcap4j.packet.IllegalRawDataException;
