+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/impl/StaticIpV4OptionFactory.java
@@ -5,10 +5,11 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.impl;
 
 import java.util.HashMap;
 import java.util.Map;
+
 import org.pcap4j.packet.IllegalIpV4Option;
 import org.pcap4j.packet.IllegalRawDataException;
 import org.pcap4j.packet.IpV4EndOfOptionList;
@@ -21,6 +22,7 @@ import org.pcap4j.packet.IpV4Rfc791SecurityOption;
 import org.pcap4j.packet.IpV4StreamIdOption;
 import org.pcap4j.packet.IpV4StrictSourceRouteOption;
 import org.pcap4j.packet.UnknownIpV4Option;
+import org.pcap4j.packet.factory.PacketFactory;
 import org.pcap4j.packet.namednumber.IpV4OptionType;
 
 /**
