+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/impl/StaticIpV6RoutingDataFactory.java
@@ -5,15 +5,17 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.impl;
 
 import java.util.HashMap;
 import java.util.Map;
+
 import org.pcap4j.packet.IllegalIpV6RoutingData;
 import org.pcap4j.packet.IllegalRawDataException;
 import org.pcap4j.packet.IpV6ExtRoutingPacket.IpV6RoutingData;
 import org.pcap4j.packet.IpV6RoutingSourceRouteData;
 import org.pcap4j.packet.UnknownIpV6RoutingData;
+import org.pcap4j.packet.factory.PacketFactory;
 import org.pcap4j.packet.namednumber.IpV6RoutingType;
 
 /**
