+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/impl/StaticSctpChunkFactory.java
@@ -5,14 +5,16 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.impl;
 
 import java.util.HashMap;
 import java.util.Map;
+
 import org.pcap4j.packet.IllegalRawDataException;
 import org.pcap4j.packet.IllegalSctpChunk;
 import org.pcap4j.packet.SctpPacket.SctpChunk;
 import org.pcap4j.packet.UnknownSctpChunk;
+import org.pcap4j.packet.factory.PacketFactory;
 import org.pcap4j.packet.namednumber.SctpChunkType;
 
 /**
