+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/impl/StaticUdpPortPacketFactory.java
@@ -5,7 +5,7 @@
   _##########################################################################
 */
 
-package org.pcap4j.packet.factory;
+package org.pcap4j.packet.factory.impl;
 
 import org.pcap4j.packet.DnsPacket;
 import org.pcap4j.packet.GtpSelector;
